requirements-devel.txt

attrs==19.3.0             # via -c requirements.txt, pytest
certifi==2020.6.20        # via -c requirements.txt, requests
chardet==3.0.4            # via -c requirements.txt, requests
coverage==5.1             # via -r requirements-devel.in, coveralls
coveralls==2.0.0          # via -r requirements-devel.in
docopt==0.6.2             # via -c requirements.txt, coveralls
factory-boy==2.12.0       # via -r requirements-devel.in
faker==4.1.1              # via factory-boy
idna==2.8                 # via -c requirements.txt, requests
importlib-metadata==1.7.0  # via -c requirements.txt, pluggy, pytest
more-itertools==8.4.0     # via pytest
packaging==20.4           # via -c requirements.txt, pytest
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via -c requirements.txt, packaging
pytest-django==3.9.0      # via -r requirements-devel.in
pytest==5.4.3             # via -r requirements-devel.in, pytest-django
python-dateutil==2.7.5    # via -c requirements.txt, faker
requests==2.21.0          # via -c requirements.txt, coveralls
six==1.15.0               # via -c requirements.txt, packaging, python-dateutil
text-unidecode==1.3       # via faker
urllib3==1.24.3           # via -c requirements.txt, requests
wcwidth==0.2.5            # via pytest
zipp==1.2.0               # via -c requirements.txt, importlib-metadata
